+++ src/main/java/br/pucrio/les/esg_token_backend/configurations/security/SecurityConfiguration.java
@@ -1,5 +1,6 @@
 package br.pucrio.les.esg_token_backend.configurations.security;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
@@ -10,10 +11,20 @@ import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import br.pucrio.les.esg_token_backend.resources.auth.services.TokenAuthenticationFilterService;
 
 @EnableWebSecurity
 @Configuration
 public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
+
+    @Autowired
+    TokenAuthenticationFilterService tokenAuthenticationFilterService;
+
+    public SecurityConfiguration(TokenAuthenticationFilterService tokenAuthenticationFilterService) {
+        this.tokenAuthenticationFilterService = tokenAuthenticationFilterService;
+    }
 
     @Override
     @Bean
@@ -30,7 +41,9 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(HttpSecurity http) throws Exception {
         http.authorizeRequests().antMatchers(HttpMethod.POST, "/auth").permitAll().anyRequest().authenticated().and()
-                .csrf().disable().sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS);
+                .csrf().disable().sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
+                .addFilterBefore(this.tokenAuthenticationFilterService, UsernamePasswordAuthenticationFilter.class);
+        ;
     }
 
     // Configuration for static resources